requirements.txt
requests
beautifulsoup4
python-dotenv
streamlit
requests
beautifulsoup4
python-dotenv
